Allow more than one cleanup handler
